pyproject.toml
[project]
name = "rainfall"
version = "0.1.0"
description = "Call functions that don't exist. Let AI figure it out."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["ai", "llm", "python", "prototyping", "gemini", "stub", "function"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "google-generativeai>=0.8.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.scripts]
rainfall = "rainfall.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/rainfall"
Repository = "https://github.com/yourusername/rainfall"
Documentation = "https://github.com/yourusername/rainfall#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rainfall"]

[tool.ruff]
line-length = 100
target-version = "py310"
